pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic"
version = "0.1.0"
description = "Orchestrator for Claude-Code-SDK agentic workflows"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "rich>=13.7",
    "claude-agent-sdk>=0.1.80",
    "textual>=0.85",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[project.scripts]
agentic = "agentic.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic"]

[tool.hatch.build.targets.wheel.shared-data]
"scaffold" = "agentic/scaffold"

[tool.pytest.ini_options]
testpaths = ["tests"]
